refactor(gastos): extraer GastoCardList compartido + reemplazar window.prompt por Modal accesible (P1 specialist)
